scripts/requirements.txt
pyrogram==2.0.106
tgcrypto==1.2.5
SQLAlchemy[asyncio]==2.0.36
asyncpg==0.30.0
python-dotenv==1.0.1
